split legacy behaviour
