+++ Packages/com.unity.inputsystem/InputSystem/LegacyInputManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 67e7b1666452a57488a40378e024a387
